+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - [TD-1660] Fix link_count to filter by business_concept-data_field relations
+- [TD-1519] Initial loader will check for missing business_concept parents relations and create them
 
 ## [2.17.0] 2019-04-17